Thread safe pool allocator

Replaced custom pool allocator with the one provided by boost
Added optional thread safety to the pool allocator using boost::mutex

See merge request !69

+++ Engine/Algorithms/TSPSolvers/SimulatedAnnealing/SimulatedAnnealingTSPSolver.cpp
@@ -106,7 +106,7 @@ namespace Scheduler
     size_t SimulatedAnnealingTSPSolver::markovChainLength(size_t stopsCount) const
     {
         const size_t M = std::nearbyint(stopsCount * std::log(stopsCount) * markov_chain_length_scale);
-        return std::max(1ul, M);
+        return std::max(size_t(1), M);
     }
 
 }
